% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBregDE2.R
\name{NBregDE2}
\alias{NBregDE2}
\title{Bayesian Negative Binomial regression for differential expression analysis of
sequencing count data with confounding factors}
\usage{
NBregDE2(counts, X, cond, idx.cond, Burnin = 1000L, Collections = 1000L,
  PGTruncation = 60L, randtry = 2017)
}
\arguments{
\item{counts}{a matrix of counts, rows are corresponding to genes and columns
are corresponding to samples.}

\item{X}{design matrix}

\item{cond}{the index of covariate corresponding to the main treatment}

\item{idx.cond}{a vector of covariate indices that the effect of their combination
on gene expression is under study}

\item{Burnin}{Number of burn-in iterations in MCMC}

\item{Collections}{Number of collected posterior samples after burn-in}

\item{PGTruncation}{the truncation level used for genrating random numbers
from Polya-Gamma distribution}

\item{randtry}{to be used for set.seed()}
}
\description{
Bayesian Negative Binomial regression for differential expression analysis of
sequencing count data with confounding factors
}
